dctest: don't chown the log files
At the end of running dctest, a recursive chown was done to make sure
gitlab-runner can remove the files that were created by root. However,
this can't work, since dctest itself is executed as user gitlab-runner,
so it is not allowed to do a chown.

Therefore, remove this. We've found other ways to make sure that the
files can be removed by gitlab-runner (by making the directories
writeable).

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>