dctest: chmod instead of chown log files
At the end of running dctest, a recursive chown was done to make sure
gitlab-runner can remove the files that were created by root. However,
this can't work, since dctest itself is executed as user gitlab-runner,
so it is not allowed to do a chown.

Therefore, do a recursive chmod a+w instead. This will allow us to
remove the files afterwards, and it will succeed since the top-level
directory is writeable by us.

Note that this is not currently needed at all, because we already made
sure that files are either removed or can at least be removed by
gitlab-runner in some other ways. However, doing this global chmod
protects us against future similar problems.

Limit this to the logs directory - that's the only directory where files
are created by root.

Do this regardless of whether the current user is gitlab-runner or not -
it is also nice when runnin tests manually that you can clean it up
afterwards.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>