periodics: add persistent_commit_data operation
This addition of persistent_data_commit_operation
class will work hand in hand with the previous
commits. The operation will check when changes
are pending to be written out to the flash -
If so it'll reset back the controlling
variable and execute a uci_commit.

Add persistent_client_data_operation
class declaration and definition.

Signed-off-by: Daniel Edri <daniel.edri@devalore.com>