agent: platform_manager: replace Socket* with int
https://jira.prplfoundation.org/browse/PPM-658

This is a preparatory commit.

Currently, controller uses `Socket *` as a handle to accepted socket
connections. Next commits will replace this with a file descriptor
which is an `int`.

To prepare for that change, rewrite all methods that get executed
whenever a socket event occurs (read and close) and create a new
method to send CMDU messages. These new methods use an `int` instead of
a `Socket *` so they can be used by the new sockets API.

Temporary methods and member variables created for this purpose will be
removed at the end of PPM-658.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>