ci: add a scheduled job to run certification tests on the isr testbed
Gitlab has no way use rule-based tags, so we have to create a separate
job for that.

We also have to change "certs-tests" to "certs-tests-essensium-1", to
make sure it gets picked up by the Essensium runner (since in the long
run we want other runners to have the "certs-tests" tag, to run MR
tests).

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>