ci: add a scheduled job to run certification tests on the isr testbed
Gitlab has no way use rule-based tags, so we have to create a separate
job.

Make the original run-certification-tests job a hidden one, and create
2 new jobs that extend it.

Add rules to run the job only if the corresponding value is set for
WFA_CERTIFICATION_RUNNER.

We also have to change the "certs-tests" tag, to make sure it gets
picked up by the right runner.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>