bpl_cfg: beerocks: adds support for uci-commit
The periodic operation interval value should
have a default constant, and a way to alter
the value on runtime.

Add cfg function for the interval, In addition
the periodic operation task id is added to
support the persistent_data_commit_operation.

Lastly, addition of missing functions to each
platform implementation unit.

Signed-off-by: Daniel Edri <daniel.edri@devalore.com>