tools: deploy_firmware: set the image name (rax40)
By default the devices do not have the image name we use set in their
u-boot environment. To avoid manual intervention and because we
already know the image name from the deploy script, we can set it from
there.

This change will also help when we change the actual image name in
prplwrt (which hopefully shouldn't happen too often).

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>