tools: deploy_firmware: set the image name (rax40)
By default the devices do not have the image name we use set in their
u-boot environment. To avoid manual intervention and because we
already know the image name from the deploy script, we can set it from
there.

Also save the name in the u-boot environment. It is not strictly
required, but has 2 advantages:

- It makes side branches which do not have this change yet usable with
any device that does not have the expected image name.

- It could (arguably) prevent confusion if a user later tries to
manually update the device without checking the name of the image in
u-boot first.

This change will also help when we change the actual image name in
prplwrt (which hopefully shouldn't happen too often).

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>