controller: move creation of tasks to constructor
Modify master_thread to create worker tasks in constructor instead of
in the `init()` method.

Worker tasks must be created only once so it is safer if they are
created in the class constructor instead of in a method that can be
called more than once. Also, if something goes wrong while creating the
tasks, application should log a fatal error and abort.

This is a preparatory commit. Next commits will replace `init()` with a
`start()` method and will add a `stop()` method to clean-up allocated
resources.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>