bwl: add dwpal context array size and use it for limit checks
Currently, the hal_dwpal uses the IFACE_TOTAL_VAPS to iterate over the
contexts array and to verify bounds before accessing the array.
This is a misuse since it is only by chance that the number of VAPs is
also the max size of the array since the first context is used for the
main radio context. if we would save in a different order this
assumption may not be relevant.

Introduce a new constexpr for max-contexts-array-size and use it instead
of IFACE_TOTAL_VAPS.
The above-mentioned constexpr is set as (max-number-of-VAPs + 1) where
the 1 is for the main radio context.

Signed-off-by: Adam Dov <adamx.dov@intel.com>