CONTRIBUTING: relax requirement on non-breaking commits
There was some discussion about the rule that a commit should never
break anything. It turns out that this rule is not that important after
all (since we've never used git bisect up to now), and in some cases it
makes our life a lot easier if a braking commit is allowed. Also, some
MRs have had breaking commits already, and there wasn't a big problem.

Therefore, relax the constraint about non-breaking commits a bit, with
an explanation why.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>