bcl: create timer manager interface and implementation
https://jira.prplfoundation.org/browse/PPM-658

This is a preparatory commit.

TimerManager and TimerFactory are intended to facilitate the creation
and destruction of timers as well as the unit testing of the classes
that use those timers.

In particular, TimerManager contains code that otherwise should have
to be duplicated each time a timer had to be created and configured.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>