bwl: common: management frame parsing
Implement a method for parsing a 802.11 management
frame, and create a notification structure.

If the given frame is not supported (errornous or not
handled by prplMesh at this point), a nullptr is returned.
Otherwise, a structure with the following fields is returned:
- STA MAC header
- Tunnelled message type
- RAW management frame body

In addition, slightly reorder the base_wlan_hal.h file by:
- Merging two "Protected methods" sections.
- Moving a protected method into the appropriate section.

Signed-off-by: Vitaly Bukhovsky <vitaly@devalore.com>