db: node: change is-friendly flag to is-unfriendly
Preparative commit.
This commit and the ones to follow change the persistent-DB
friendliness parameter to indicate unfriendliness. This is done since
the clients are treated as friendly unless states otherwise, hence it
is more proper to indicate in the persistent DB unfriendliness as the
parameter is not configured in the persistent DB, it is treated
as friendly.

Changed the field in the client node from is_friendly to is_unfriendly.
The default value is not-configured which is treated as false.

The commits separation breaks the build and is meant for better
readability and easy review.

Signed-off-by: Adam Dov <adamx.dov@intel.com>