aging_task: change the is-friendly flag to is-unfriendly
Change the persistent-DB aging task to use the is_unfriendly parameter
logic instead of the is_friendly.

If a client is not set explicitly as not-friendly in the persistent DB,
it is treated as a friendly client.
Friendly clients use the global max-time-life-delay parameter for
aging, while unfriendly clients use the
unfriendly-device-max-time-life-delay parameter (which can be set to a
lower value).

Signed-off-by: Adam Dov <adamx.dov@intel.com>