steering_task: change the is-friendly flag to is-unfriendly
Set the unfriendliness of a client to true only if that client is
already part of the persistent DB.
This reason for this change is that unfriendliness has currently no
meaning for clients without persistent information.
We don't want to fill the persistent DB with clients that only have the
unfriendliness parameter configured.

Signed-off-by: Adam Dov <adamx.dov@intel.com>