agent: autoconfig: put WSC TLV at the end of WSC CMDU
With the addition of the R2 TLVs to the WSC CMDU, the Marvell controller
no longer responds to the M1 message, which makes e.g.
MAP-4.2.1:axepoint fail.

As a workaround, put the WSC TLV at the end of the M1 message, i.e. let
the other TLVs precede the WSC TLV. This is probably a bug in the
Marvell controller, but the order of the TLVs doesn't matter to us and
this way we can still pass certification.

Fixes: PPM-698

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>