common: Add 5G channels table
Preparative commit.
Add 5G channels table which contains for each channel the all supported bandwidths
with its center channel and the min and max overlapping channels range.

This is table is very helpful and could be used for any channel selection purpose.
Specifically for ZWDFS feature, this table is required for selecting the channels with
higher bandwidth which has the same rank as lower bandwidth channel.
If a channel has a bandwidth greater than 20Mhz, need to check that its overlapping
beacon channels are usable/available too before selection this channel.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>