agent: Add a socket argument to agent task handle_cmdu
Add a Socket argument to the task infrastructure on the agent, in "handle_cmdu" functions.
Update all required handlers on each of the Agent tasks.

This change is required since the channel selection task on the agent needs it
for understanding which radio has sent the message since the system flow is different on
each radio interface that sends a message.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>