agent: Add channel selection task friend of the backhaul manager
Preparative commit.
Channel selection will need to send messages to some radios.
Since the tasks do not hold any sockets, they need to use the sockets which are
held on the thread running the task, which is currently the backhaul manager.
Since the  sockets container in the backhaul is a private member, add the
channel selection task as a friend of the backhaul manager.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>