agent: Add ZWDFS interface related member variables and function
Add to the Agent channel selection task two members variables.
The one holds the interface of the primary 5G radio, and the other for holding the
ZWDFS radio interface.
In addition, add a function that is responsible to initialize the ZWDFS interface
member on the task, and shall be called when initiating the ZWDFS flow on the task.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>