agent: Add zwdfs_select_best_usable_channel function
Add the function "zwdfs_select_best_usable_channel" to the channel selection task.
The function is reponsible to calculate what is best available/usable channel from
the channel list for the ZWDFS flow.
The function iterate on all the channels from the channels list on the AgentDB, and
lookup for the channel with the lowest rank (lower=better). If a channel with
bandwidth higher than 20MHz has the same rank as the 20MHz channel, the
function will select it only if all of the overlapping beacon channels are
available/usable too.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>