cppcheck: Update cppcheck existing issue file
Since it is a preparative PR, there are new functions that are not used currently but will be on the follow-up PR.
Unused functions raise a cppcheck issue.
Update the cppcheck existing issue so it will not block the merge.
The new cppcheck issue will be resolved on the next ZWDFS PR.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>