nbapi: add parameter argument to Ambiorix::set()
The Ambiorix::set() family of functions uses the relative path to
determine the parameter to set. However, the underlying amxd functions
expect the path to the owning object for the transaction, and only the
parameter name for setting it.

We could parse out the parameter name from the relative_path, but it's
much simpler to explicitly add the parameter as an additional argument
to the set() call.

Add parameter argument to the set() function, and update all callers
accordingly.

While we're at it, also fix the callers to include the "Controller."
part of the path. It turns out that the Controller object is not the
root, the root is an unnamed object above it.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>