db: return false for is-client-in-persistent-db if disabled
If the persistent DB feature is disabled, and the
is_client_in_persistent_db() is called - the API should return false
instead of using the BPL to perform the check.

Add check of persistent DB to is_client_in_persistent_db() API and
return false if the feature is disabled.

Signed-off-by: Adam Dov <adamx.dov@intel.com>