ci: use merge requests label to trigger certification tests
When on a merge request, mentioning the tests and targets in commit
description only works if the MR branch is up-to-date with
master (which is rarely the case).

Switch to a new workflow:

The target has to be mentioned in merge request labels. To trigger the
builds/tests on the very first pipeline of the merge request, the
labels have to be added when the MR is created. This can be done using
Gitlab quick actions [1]. For example, adding the following line to a
merge request description would add the labels
'certification:MAP-4.2.1' and 'target:turris-omnia':

/label ~certification:MAP-4.2.1 ~target:turris-omnia

The previous way of triggering certification tests from commit
messages is retained, because it can be useful when used on
development branches (for which we don't intend to create merge
requests).

To achieve this, first the builds for Turris Omnia and Axepoint now
have an additional rule to be built if the merge request labels
contain "target:" followed by the target name.

The wired backhaul tests are generated using the following snippet:
target=turris-omnia
cat ../../tests/certification/passing_agent_tests.txt | while read -r i ; do
printf '%s:%s:\n  extends: .certification:%s\n' "$i" "$target" "$target" >> "$target.yml"
printf "  rules:
    # If the last commit description contains the test name and we're not on master, run it.
    # Otherwise, make it a manual job
    - if: '\$CI_COMMIT_DESCRIPTION =~ /.*%s:%s.*/ && \$CI_COMMIT_REF_NAME != \"master\"'
      when: on_success
    # If we're on a MR and the labels have the correct target and test name, run it.
    - if: '\$CI_MERGE_REQUEST_LABELS =~ /.*target:$target.*/ && \$CI_MERGE_REQUEST_LABELS =~ /.*certification:$i.*/'
    - when: manual
      allow_failure: true\n
" "$i" "$target" >> "$target.yml"
done

The wireless backhaul tests are generated in a similar way, using the
following snippet:
target="turris-omnia"
grep _BH ../../tests/certification/all_agent_tests.txt | while read -r i ; do
  printf '%s:%s:\n  extends: .certification:%s\n' "$i" "$target" "${target}-bhwifi" >> "${target}-bhwifi.yml"
  printf "  rules:
    # If the last commit description contains the test name and we're not on master, run it.
    # Otherwise, make it a manual job
    - if: '\$CI_COMMIT_DESCRIPTION =~ /.*%s:%s.*/ && \$CI_COMMIT_REF_NAME != \"master\"'
      when: on_success
    # If we're on a MR and the labels have the correct target and test name, run it.
    - if: '\$CI_MERGE_REQUEST_LABELS =~ /.*target:$target.*/ && \$CI_MERGE_REQUEST_LABELS =~ /.*certification:$i.*/'
    - when: manual
      allow_failure: true\n\n" "$i" "$target" >> "${target}-bhwifi.yml"
done

[1]: https://docs.gitlab.com/ee/user/project/quick_actions.html

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>