ci: make run-certification-tests usable with turris-omnia
In the long run we probably want to switch to using separate jobs for
each test to avoid having to repeat the same script twice (it's also
in ci/certification/). This is however not possible yet because we
need the google sheet upload (done in run_test_file.py) to be done for
all the tests at once.

For now, change run-certification-tests to make it usable with the
Turris Omnia.

Change the check for deploy_firmware to include the turris-omnia.

Also add the "build-for-turris-omnia" job in "needs", to fetch the
artifacts for it.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>