bcl: remove offset parameter from Connection::receive()
https://jira.prplfoundation.org/browse/PPM-658

This is a preparatory commit.

The `offset` parameter in `Connection::receive()` is never used.
Moreover, the `receive()` system call must append received data to
existing buffer contents, if any (this is what the `offset` parameter
was initially intended for, but later we added the `length()` method).

This fix is necessary in stream-oriented sockets, to deal with the case
where a message is not received at once, but fragmented.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>