agent: backhaul manager: create channel selection task
Backhaul manager creates a bunch of tasks.
When a new task needed to be introduced the code of
an old task was copy-pasted and edited.

This approach is error prone and once it failed.
Instead of channel seletion task the second
autoconfiguration task has been created.

This commit creates fixes this copy-paste mistake.

Signed-off-by: Vitalii Komisarenko <vitalii.komisarenko@gmail.com>