transport: fix bug on socket disconnected
When a client socket gets disconnected, the broker server calls back the
`socket_disconnected()` method. This method must erase the socket from
the list of subscribers to a given message type.

Method `socket_disconnected()` is accepting a shared_ptr, this pointer is
captured by a lambda and this lambda is stored as part of the event
handlers registered for the socket. When the socket gets disconnected,
the event loop is automatically removing event handlers for the socket,
the lambda is not stored anymore. As a result, shared_ptr passed to
`socket_disconnected()` contains a dangling reference so the method does
not remove the subscriptions for the disconnected socket. As a side-
effect, when broker tries to publish a message to the disconnected
socket, the whole transport process crashes.

The solution has been not to store a shared_ptr in the lambda function
but a file descriptor and keep a map of fd to shared_ptr as a class
member variable.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>