transport: erase socket from subscribers on disconnect
When a client socket gets disconnected, the broker server calls back the
`socket_disconnected()` method. This method must erase the socket from
the list of subscribers to a given message type.

However, the socket was being erased from a copy of the subscribers, not
from the list of subscribers itself. To fix this, we must get a
reference to such list of subscribers (instead of a copy) and then erase.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>