dwpal: During STA_CONNECT generation skip unassociated clients
Preparative Commit

In the prplMesh startup process, there's a point at which it'll generate STA_CONNECT EVENT for clients that were previously connected. As a general rule of thumb, clients that are not associated appear to lack the data field 'connected_time' in their run-time configurations and hence shouldn't change their connection state - the solution we purpose in this commit allows skipping any clients that are not already associated.

adds an enum 'generate_association_event' to elaborate on what's been going on in the generate_client_assoc_event flow during run time.

Upcoming changes are regarding the alteration of 'generate_client_assoc_event'
and generate_connected_clients_events.

Signed-off-by: Daniel Edri <daniel.edri@devalore.com>