bwl: dwpal: generate STA_CONNECT event only for associated clients
Preparative Commit.
On prplmesh init we generate STA_CONNECT event for
already connected clients. Currently, we do this for
all clients listed on the STA-FIRST/STA-NEXT replies,
but actually, the above commands will return also
clients that are authenticated but not yet associated.

To generate the event only for associated clients we
can read the connected_time field which exists only
for associated clients and if not exist not trigger
the event.

Add enum to extend the possible outcomes of
generate_client_assoc_event() to also return
"client-is-not-associated" which will skip
generating the event for that client.

PPM-371

Signed-off-by: Daniel Edri <daniel.edri@devalore.com>