bcl: EventLoopImpl: keep handlers by value
In the main event loop, the handlers are looked up in
m_fd_to_event_handlers and then called depending on which event(s)
happened. Since the handlers is a struct of four std::Function members,
it is certainly larger than a single word so it is taken by const-ref.

However, in the error and HUP cases, remove_handlers() is called before
calling the event handler itself. remove_handlers() removes the handlers
from m_fd_to_event_handlers, thus it destroys the handlers object.
Therefore we're accessing already-freed memory when calling the handler.

To fix this, take a copy of the handlers instead of using const-ref.
In practice, the compiler will most likely optimise away the full copy
and instead copy only the single handler that is going to be called, so
the impact should be minimal.

The alternative would be to first call the handler and then
remove_handlers. This has three disadvantages:
- it's easy to forget about this and re-introduce the issue later;
- if the handler itself does remove_handlers as well (which is an
  error), the problem will be reported in the context of the run()
  function instead of where it actually is;
- the handler will probably close the fd, which implicitly removes it
  from the epoll - if we explicitly remove it afterwards, epoll_ctl will
  report an error.

Fixes PPM-714

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>