ci: make 4.2.1 run on all merge requests
Now that 4.2.1 is passing again and that we're about to enable a
second testbed in CI, we can afford running it for every merge train
pipeline.

Also add a "retry" of 2 for this specific test, because we know the
certification tests are unstable, and we don't want a merge train
pipeline to fail because of an unstable test.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>