son: Alters client_get_client request to also send time_life
The son manager is the bridge between the BML internal
and the DB. For our purpose of extending this client
requests to send the max_time_life, we need to modify
this request.

Alters client_get_client to be concious of
time_delayed_minutes.
Remove conditions which checks for UNCONFIGURED_ARGUMENT.

Signed-off-by: Daniel Edri <daniel.edri@devalore.com>