bcl: create bridge status manager interface and implementation
https://jira.prplfoundation.org/browse/PPM-750

Create a bridge status manager component to read the interfaces
included in a network bridge as well as to get notified with an event
whenever either a new network interface is added to the bridge or
removed from the bridge.

The bridge status manager will be used in comming merge requests.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>