bcl: create bridge state manager interface and implementation
https://jira.prplfoundation.org/browse/PPM-750

Create a bridge state manager component to read the interfaces
included in a network bridge as well as to get notified with an event
whenever either a new network interface is added to the bridge or
removed from the bridge.

The bridge state manager will be used in future merge requests.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>