gitlab-ci: split off platform jobs in a separate file
Split off the jobs related to building on platform into a separate file:
ci/platform.yml. Certification tests that run on a platform are also
moved to this file, but the dummy certification test is kept at
top-level.

This makes no functional difference, it just reorganises code.

While we're at it, put the include blocks near the beginning of the
file, which is a more usual place for include-type statements. Note that
the yaml file is declarative, so the location makes no difference.

In a similar vein, move the upload-artifacts job to the end of the file,
which is more logical since it's the last stage.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>