gitlab-ci: move all templates to ci/templates.yml
When we will create child pipelines in a later commit, the templates
that currently exist in gitlab-ci.yml will sometimes be used by the
top-level pipeline, sometimes by the child pipeline, and sometimes by
both.

To ease the transition and to reduce the cognitive load, move all
templates to a single file, ci/templates.yml, which can be included both
in the top-level and in the child pipelines.

.rules-for-expensive-job is a special case: it is not really a template,
it's more like a macro. Unlike a real template, it gets expanded by yaml
itself, not specifically for gitlab-ci. Since the include statements
*do* get expanded by gitlab-ci, we can't define rules-for-expensive-job
in one file and use it in another file. And since this macro gets used
by some templates *and* directly in some jobs, we need to define it in
both files.

This duplication will be removed again in a later commit.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>