gitlab-ci: execute platform jobs as a child pipeline
Moving these jobs to a child pipeline has two advantages:
- The pipeline page is cleaner (especially because the certification
  tests are removed).
- It is possible to manually trigger all the expensive jobs in the child
  pipeline with a single click.

Note that the latter is not possible yet in this commit, because the
jobs in the child pipeline still have the 'when: manual' rule. A
follow-up commit will fix that.

A pipeline can only refer to jobs within the same pipeline in its
"needs" specification. Therefore, references to the build-on-[platform]
needs clauses (for the certification template and for the upload job)
are replaced by a reference to the platform job.

The platform job has a 'strategy: depend' to make sure it takes over the
status of the child pipeline. Without this, merely creating the child
pipeline would be sufficient to consider the job to pass.

The run-certification-tests and upload-artifacts jobs need artifacts
from the child pipeline. The only way to get those is to use the
"Artifact downloads between pipelines in the same project" approach.
Unfortunately, this is not perfect, because it uses the _latest_
pipeline that corresponds to the given ref, which is not necessarily the
child pipeline corresponding to the current one (if a concurrent
pipeline was triggered on the same ref). Also it's not clear what will
happen in merge train pipelines - but these jobs are anyway not
triggered for merge trains.

The child pipeline needs its own workflow, otherwise it apparently has a
workflow that doesn't match anything.

For more documentation about child pipelines, refer to
https://docs.gitlab.com/ee/ci/parent_child_pipelines.html

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>