agent: ap_autoconfiguration_task: increase number of attempts
When configuration task exceeds the maximum number of attempts
 for configuration search backhaul manager goes to reset.
In case of no response backhaul manager reset accrues every 20 sec.
In some cases it takes longer than that for a response to arrive
and it could arrive in the middle of the reset phase.
Increase number of attempts and timeout between attempts
to overall 180sec.

Signed-off-by: Alex Kanter <alex.kanter@devalore.com>