transport: add the bridge state manager to the transport.
This is a preparatory commit.

The bridge state manager is added to the transport, but it's not used
for anything yet. It is added in a similar way to how the interface
state manager is created.

Also note that the registered handler will not be called yet, because
"read_state()" has to be called once on the manager to make it start
monitoring the bridge. This will be done in a follow-up commit.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>