transport: add the bridge status manager to the transport.
This is a preparatory commit.

These changes are taken literally from the commit adding the interface
state manager (08915599ce917d0028df20db8316e5322591e717), with a few
changes specific to the bridge status manager.

The bridge status manager is added to the transport, but it's not used
for anything yet.

Also note that the registered handler will not be called yet, because
"read_status()" has to be called once on the manager to make it start
monitoring the bridge. This will be done in a follow-up commit.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>