controller:odl: add mib for ht(vht/he)capabilities
If a Data Elements radio is HT capable, we shall include the Capabilities.HTCapabilities object.
If a Data Elements radio is VHT capable, we shall include the Capabilities.VHTCapabilities object.
If a Data Elements radio is HE capable, we shall include the Capabilities.HECapabilities object.
So in some cases, with some conditions we need to add extra singleton object(s).

Made mid for HECapabilities, VHTCapabilities and HTCapabilities
as we will add them to 'Capabilities' object only when it's necessary.
Make Capabilities not a multi-instance object.

Signed-off-by: Krystyna Oliinyk <k.oliinyk@inango-systems.com>