use the bridge state manager to reconfigure the transport
btl_broker:

- configure_ieee1905_transport_interfaces(): fill the configuration
request with the bridge only. It previously included a list of all the
interfaces, but that will now be done by the bridge state
manager. There are a lot of things in the configuration message that
are now unused and can be removed, but it requires more refactoring,
and thus is not done as part of this commit.

- handle_broker_interface_configuration_request_message(): instead of
taking the list of interfaces from the configuration message, use the
bridge manager's read_state() to get them.

- handle_bridge_state_change(): whenever a change in the bridge occur,
we know which single interface was involved, so we can use it to
add/remove it to the transport.

Fixes https://jira.prplfoundation.org/browse/PPM-749

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>