framework:nbapi: create add/remove subobject methods
Create add_optional_subobject() method, which will instantiate optional sub-object.
The subobject must be defined as a mib in the odl file. The name of the mib must be
the same as the name of the subobject, and it must contain only a single object definition.

Create remove_optional_subobject() method which will remove optional subobject
from specified object.

PPM-482
--
Signed-off-by: Krystyna Oliinyk <k.oliinyk@inango-systems.com>