bcl: create netlink event listener interface and implementation
https://jira.prplfoundation.org/browse/PPM-750

This is a preparatory commit.

The goal of PPM-750 is to create a component to monitor the network
bridge and get notified whenever a new interface is added to the bridge
or removed from the bridge.

Information about bridge changes can be obtained by parsing Netlink
events received through a Netlink socket connection, pretty much the
same as like we are doing in the InterfaceStateMonitor to get notified
about changes in the up and running state of network interfaces.

To avoid duplicating code and to avoid creating another Netlink socket
for this purpose (thus providing a scalable solution), a new
NetlinkEventListener component is created in this commit. This
component will include all the code to manage the Netlink socket
connection and to parse received messages. Later commits will include
the NetlinkEventListener in the InterfaceStateMonitor and in the
BridgeStateMonitor, which will register handlers to get notified of
Netlink events and parse them accordingly.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>