bcl: use netlink event listener from interface state monitor
https://jira.prplfoundation.org/browse/PPM-750

Modify InterfaceSateMonitor to replace the code to manage the Netlink
socket connection and the event loop with the brand new class
NetlinkEventListener.

Modify the main() function in the transport process to create the
NetlinkEventListener and pass it to the InterfaceStateMonitor. If a
BridgeStateMonitor is created in the future, the very same instance of
NetlinkEventListener will be used.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>