dev: boardfarm: Changed _run_shell_cmd function
Function signature:
It makes sense that the default value of a list is set to None,
since we could be mutating the list inside the function
making it spill the value  when the function is re-used.
However, this is not the case, because no value is being returned,
so there's no chance to stack whatever value it holds.
On the bad side, pexpect.run has a inner validation checking if
args is an instance of type list, which would make an exception if
we ever decided not to fill this argument.

Removing env from pexpect.run call:
Pexpect run function already defaults the env param to None, so
making a pre-condition to call it with and without the env is
unnecessary redundancy.

Changing the command structure:
Previously if we sent a pipe on a command, this would be interpreted
by the printf installed on /usr/bin/printf, instead of the bash
builtin version. This would cause the extra arguments to be ignored
in a specific case, when running `printf device_get_info | nc ...`
all the extra arguments were being ignored, but since the exit
status code of the printf is still 0 (it was a warning, not an
error), no flags were raised saying the command was not doing
anything.
By making the command run with /bin/bash, we can assure the use
of the right interpreter.

Final Note:
Using a command is list form (ex: ['printf', 'this'] does look
more clean and is easier to maintain between OS's, this might
not be the final version

Signed-off-by: David <davidmgafonso@gmail.com>