dev: boardfarm: Changed check_status function in prplmesh_compose
Changed the call to _run_shell_cmd:
The call was changed to better fit the changes specified on
the prplmesh_base commit.

Added a get_active_entity function:
In the future, to avoid constantly checking whether a device
is supposed to be an agent or a controller, thought it was a
good idea to make a function that just returns the one that
was chosen on instantiation, since there can/should only be
one anyway.
This can be removed if proven redundant.

Signed-off-by: David <davidmgafonso@gmail.com>