dev: boardfarm: Changed prplmesh_prplwrt check_status function
Changed the call on check_status to use beerocks_cli:
The nc command on the board does not seem to support
the -w flag, even when running nc -h to potentially
check the version, the output is different from my
local env.

Change the call to sendline on get_prplMesh_status:
There's a function inplace to abstract the call to
/etc/init.d/prplmesh, so maybe this call should be
standardized.

Signed-off-by: David <davidmgafonso@gmail.com>