dctest: fix using the boardfarm image built in the pipeline
In 2630d77b7fbc416dff90df4dfc17f41f97e2512f, the condition to set
CI_PIPELINE_ID to "latest" only if running locally was removed. As a
consequence, the pipelines were always using the latest image.

Add it back, to make the CI pipelines use the correct boardfarm image.

'local_run' was also part of the condition that was removed but it
turns out it's unused, so remove it as well.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>