controller: remove active fail-safe channel switching from channel selection
Fail-safe channel settings needs to be handled by the hostapd, but currently
the controller channel selection task monitors CSA/CAC and sets the fail-safe
channel according to some logic.
Remove any controller logic that may actively switch to fail-safe
channel, but keep the db settings so it can be reused or removed completely
in the future.
Done as part of zwdfs feature changes PPM-630.

fix cppcheck issue as well.

Signed-off-by: Alex Kanter <alex.kanter@devalore.com>