bcl: add TcpServerSocket::create_instance
https://jira.prplfoundation.org/browse/PPM-658

Currently we have a method to create an instance of a TCP server
socket in beerocks_master_main.cpp. In the future, this method will
also have to be added to beerocks_slave_main.cpp

To avoid duplicated code, move such method to a new common factory
method in a new class TcpServerSocket.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>