bcl: add UccServerFactory::create_instance
https://jira.prplfoundation.org/browse/PPM-658

Currently we have code to create an instance of a UCC server socket in
beerocks_master_main.cpp. In the future, this method will also have to
be added to beerocks_slave_main.cpp

To avoid duplicated code, move such code to a new common factory
method in a new class UccServerFactory.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>