bcl: group all three CMDU server handlers in a struct
https://jira.prplfoundation.org/browse/PPM-773

CMDU server defines three event handlers to deal with on connect, on
disconnect and on CMDU received events. Since the three handlers are
always set together, it would be convenient to have a single function
to set all three of them at once, similar to how it is done in the
EventLoop class.

Refactor CMDU server to group all handlers in a struct and change
the methods to set and clear the handlers. Also refactor platform
manager and controller to use the new methods.

While we are at it, add an on-connect event handler to the CMDU server
defined in platform manager so a message is written to log when a
client connects.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>